refactor(sheet): Remove legacy InventoryPanel and relocate shared components (#380)
Remove the legacy InventoryPanel (superseded by inline controls in individual
display components) and relocate WirelessIndicator and WeaponAmmoDisplay to
components/character/sheet/ where they belong with their consumers.
