+++ components/character/sheet/AugmentationsDisplay.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import type { Character, CyberwareItem, BiowareItem } from "@/lib/types";
 import { isGlobalWirelessEnabled } from "@/lib/rules/wireless";
 import { DisplayCard } from "./DisplayCard";
-import { WirelessIndicator } from "@/app/characters/[id]/components/WirelessIndicator";
+import { WirelessIndicator } from "./WirelessIndicator";
 import { ChevronDown, ChevronRight, Cpu, Wifi, WifiOff } from "lucide-react";
 
 // ---------------------------------------------------------------------------
