+++ components/character/sheet/WirelessIndicator.tsx
